Run the trial.html for the homepage.
This will be my final build for this project. Since I was restricted to using only HTML and basic CSS, I couldn't get anywhere close to the website that I will make for myself in the future. But, I have given my best to complete this. I wanted to customize it more, but I was not able to figure out some parts.
Hope you like it!!